BRD-1770 調整GetOrderInsuranceInfo：增加5個output欄位1是否有共同承租人2機車安心服務基本分鐘3機車安心服務基消金額4機車安心服務分鐘(幾分鐘算一次錢)5機車安心服務金額(一次算多少錢)

+++ WebAPI/Controllers/GetMotorRentProjectController.cs
@@ -241,7 +241,7 @@ namespace WebAPI.Controllers
                                 CarType = lstData[0].CarType,
                                 CarTypeName = lstData[0].CarBrend + ' ' + lstData[0].CarTypeName,
                                 CarTypePic = lstData[0].CarTypePic,
-                                Insurance = lstData[0].InsuranceDef,
+                                Insurance = (int)lstData[0].InsuranceDef,
                                 InsurancePerHour = 0,
                                 IsMinimum = isMin,
                                 Operator = lstData[0].Operator,
